requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
groq
python-dotenv
rich
pandas
beautifulsoup4
lxml
httpx
